Get service to work with bucket and key info
Changes path variables to query parameters so that file path structure
of S3 object keys can be understood. If there is a way to use the MD5
hash of these objects to retrieve, it would look better. But I don't
believe that is an option, at least not in the Boto library.

Everything else are misc refactors.