Fix multiple bugs and implement "video search"
Implement video search means that I added a REST endpoint for /video,
so the user can search for a video S3 object key. The only real
difference is that it will search for the audio S3 object key in our
database and use the audio key instead.

Fixing multiple bugs means the following:

- Fix quotations around Python script arguments to account for spaces
- Fix Python script execution so error is thrown when it fails
- Fix missing space in argument of Python script